Fix up and simplify ordering of load a bit more
