+++ loader/src/main/java/net/neoforged/fml/loading/mixin/FMLMixinTransformationService.java
@@ -13,22 +13,19 @@ import java.util.List;
 import java.util.Set;
 
 public class FMLMixinTransformationService implements ITransformationService {
-    private MixinFacade facade;
-
     @Override
     public String name() {
         return FMLMixinLaunchPlugin.NAME;
     }
 
     @Override
-    public void initialize(IEnvironment environment) {
-        var plugin = environment.findLaunchPlugin(FMLMixinLaunchPlugin.NAME).orElseThrow(() -> new IllegalStateException("FMLMixinLaunchPlugin not found!"));
-        this.facade = ((FMLMixinLaunchPlugin) plugin).getFacade();
-    }
+    public void initialize(IEnvironment environment) {}
 
     @Override
     public List<Resource> beginScanning(IEnvironment environment) {
-        return List.of(new Resource(IModuleLayerManager.Layer.SERVICE, List.of(facade.createGeneratedCodeContainer())));
+        var plugin = (FMLMixinLaunchPlugin) environment.findLaunchPlugin(FMLMixinLaunchPlugin.NAME).orElseThrow(() -> new IllegalStateException("FMLMixinLaunchPlugin not found!"));
+
+        return List.of(new Resource(IModuleLayerManager.Layer.SERVICE, List.of(plugin.createGeneratedCodeContainer())));
     }
 
     @Override
